Add cuDNN implementation of the neural network
Implement the neural network in the Rust engine using cuDNN and add a
simple self-play module without any search. The engine is still
relatively weak but wins against simple engines such as GNU Go.

The self-play module uses the neural network to predict each move
together with some dirichlet noise using a mixing ratio of 90% policy
and 10% noise. It plays randomly weighted by the prior value supplied by
the policy for the first 10 moves of the game. After the 10 first moves
the rest of the game is played out in a greedy manner.

This results in some rather funny openings, and sometimes loopsided
games but once the engine is past the opening it starts performing
respectible for someone playing only on intuition. The value output from
the network is only used to determine if the engine should resign.