Fix cuDNN implementation and add a simple self-play module
Fix the following two bugs in the cuDNN implementation toghether with a
bunch of improvements (such as async streams to allow for computing the
policy and value head in parallel):

* The policy filter was not transposed correctly, resulting in garbage output.
* The matrix multiplication arguments were wrong.

Also add a self-play module that uses the neural network to predict each
move (with no search) together with some dirichlet noise. The first 10
moves are randomly picked with the prior value suggested by the neural
network as the weight, after those 10 moves the rest of the game is
played out greedily.

This results in some rather funny openings, and sometimes loopsided
games but once the engine is past the opening then it starts performing
respectably even if it mis-reads life and death sometimes as one would
expect from a pure policy network.

The value output from the network is only used to determine when to
resign, which the engine does once the value hits below -0.9. This works
well with a human judge but might not work too well with TT scoring
since most games ends up a huge group of dead stones still on the board
in a non-trivial capture race.