Simple MCTS
Implemented a simple monte-carlo tree search with neither parallel
computation nor RAVE. The UCB formula is the same as the one used in the
AlphaGo Zero article. The tree search always perform exactly 2,000
probes into the tree.

The tree-search version of the engine wins against the pure policy 30
out of 30 attempts (so far), usually by killing some group of the pure
policy engine.