Parallel MCTS implementation
Make the monte carlo search tree algorithm run in 16 parallel threads
with a neural network batch size of 8. This code features a custom
spin-lock for handling parallel updates of the search tree, reads are
not protected by any locks how due to the fact that all reads are
aligned there should be no corrupted reads.

The engine performs about 30% faster with a parallel implementation,
this number is smaller than 1600% (the number of threads) because it
bottlenecks hard on the neural network evaluation.

As a sanity check, the engine still wins 12-0 against the pure policy
network player.