Add a tracer to the MCTS that dumps the entire search tree as an SGF
Add the compile-time feature `trace-mcts` that dumps the entire search
tree (after every move) to standard output as an SGF file. This is
useful for debugging or monitoring the behaviour of the monte carlo
algorithm.

This patch also fixes the following weird behaviours that were detected
due to this tracer:

* Do not allow duplicate expansions of the same node (instead force
different nodes at the leaf level).
* Symmetry breaking in the prior moves (by setting one of the symmetric
moves to -Inf), this is mostly useful in the opening to avoid duplicat
efforts.

These changes improves the strenght significantly, in the statistics
below the implementation with these fixes are denoted as `v2`, the other
ones are without these fixes. All of these tests were ran with 500
iterations:

* PUCT vs PUCT: *4 vs 16*
* PUCTv2 vs PUCTv2: *8 - 12*
* PUCTv2 vs PUCT: *9 vs 11*

As you can observe Black received a significant strength increase, going
from 20% win rate with the old algorithm to 40% win rate with the fixes.