Add tool for unbiasing a _big SGF_ file
Add the `tools/sgf2balance.py` script that reads SGF files from standard
input and outputs an a sub-set of the given games, such that the
outputted set of games is perfectly balanced between black and white
wins.

This is useful because many dataset, for example the KGS data-set, is
heavily biases which screw the training of the neural network.