Experimental support for FP16 in `nn`
Add experimental support for half precision in the neural network parts
of the code. So far this does not fully work, because of what I suspect
are lacking hardware and drivers on my part:

* The batch normalization throws a `BadParam` error. This is probably
  due to me using CUDA v8.
* The FP16 code path is really slow, this is expected since my GPU
  does not have true half precision support.

An alternative approach would be to use INT8 in `nn` instead of FP16, which
would work better on almost all commodity hardware than FP16 (which is
really only fast on NVIDIA's data centre cards, most notably the V100).