Add some benchmarks for comparing i8, f16, and f32 performance
Add six benchmarks for testing the performance of cuDNN when operating
in `i8`, `f16`, and `f32` mode over batch sizes 256 and 128. The `f16`
number will be huge and not accurate for _real_ compute cards since the
card I am using does not have true `f16` support:

* i8: 52423 ns / 30827 ns
* f16: 13246299 ns / 4422870 ns
* f32: 72464 / 22045 ns

Suprisingly `i8` only shows a 30% improvement over `f32` (probably since
it uses an `i32` accumulator internally), and `f32` is even faster when
running on a smaller batch size. This suggests that moving inference to
`i8` is not a viable option, and we should instead focus on either `f32`
or `f16` inference depending on the _Compute Capability_ of the card.