Fix `f16` inference code path
Change the parameters used for the batch normalization to always be
`f32` regardless of the type of the input and output tensors. This fixes
the `BadParam` crash that batch normalization were showing before when
using `f16`.

Also make sure that the relavant weights are always loaded as `f32`,
this is accomplished by checking if the name contains "/mean" or
"/variance" at the moment. A better solution would be desirable.