Generate features in the correct data type
Previously the board features would always get generated in `f32`, and
then converted to `f16` (if necessary) on the inference thread. This is
a bit of waste of resources, since any time the inference spends not
doing inference is wasted time.

This patch removes the code for converting to and from `f32` from the
neural network package and makes it the monte carlo search threads
responsibility to send it the correct type instead. This has two
advantages:

* We can generate the features directly in `f16` and skip the conversion
  step completely.
* Off-load some work from the inference thread, allowing it to perform
  more inference instead.