Add support for half precision in the inference path (#2)
Add support for half precision and tensor cores in the `nn` package

In order to support modern architectures add support for half
precision (f16) and tensor cores in the `nn` package. This requires us
to re-factor a lot of the code that generates the features and how we
store the weights internally, since those needs to be in half precision
too for this to work. This is mostly solved using generics, which
results in a fair amount of duplicate machine code and a larger binary
size but has no real runtime impact.

Most notably the batch normalization layers still expect their
parameters (mean, variance, scale, and bias) to be stored in single
precision mode. This currently has a bit of a hackish solution.

See the _Hardware Benchmark_ wiki page for a performance comparison when
running with support for half precision on a V100 compared to not (about
a 2x speed-up).

This patch also contains some bug fixes that were discovered in the
process. Most of them are related to numerical imprecision that
more easily arise with half precision.