Decrease the number of neural network channels to 128
As discussed on the "Tuning ~ Neural Network" wiki page decreasing the
number of channels in the neural network gives about a 3x speed-up with
no observable loss of strength.

Based on a 20 game tournament between an 256 channel and a 128 channel
version of _Dream Go_ the final score was 11-9 (in favour of the 128
channel version). This is within the margin for error and they are as
far as we can tell equivalent.