Change how the MCTS handle double-expansion
Because of the fact that we expand the tree with every probe into the
search tree, we cannot have two threads follow the exact same path.
Previously this issue was solved by forcing all threads to take
different routes down the search tree, but this is bad for two reasons:

1. It does not _obey_ the UCT value, so it no longer guarantees
   convergence to mini-max.
2. When there are fewer valid moves than the number of threads, some of
   them will start exploring invalid moves. This waste evaluations
   causing a weaker end-game play.

This patch changes the behaviour so that when a thread detects that the
choosen path is already taken, it enters a _paused_ state where it waits
until the tree has been changed before it tries another probe into the
search tree.

This required us to add support for partial batches in the code as there
may be fewed than the batch size valid moves (especially in shallow
search trees during the end-game).

This fix makes the engine significantly stronger, winning 64% to 74% of
its games against the version without this fix over 20 games.