Neural network performance optimization
Tweak the implementation of the neural network and how the monte carlo
tree search handle batching of requests to improve efficiency by about
5% - 15%.

* Re-write the `waiting` queue to use channels and directly tell the
  server thread that it is waiting, instead of relying on polling of
  an atomic counter.
* Rename the previous `Network` to `Shared` and introduce a new
  `Network` structure that is a workspace pool.
* Use workspaces of different sizes more aggresively to maximize the
  batch size.

These optimizations are necessary because of how we changed probing into
the monte carlo search tree to not force divergent paths, and instead
fail evaluations. Causing non-uniform batch sizes, especially during the
beginning of the search.