Tweak the neural network batch size and the number of parallel threads
Change the default values for the neural network batch size to 16 and
the number of threads to 64. This is motivated by a desire for a batch
size that is as large as possible, and enough threads to keep the batch
saturated. See the wiki page "Tuning ~ Neural Network" for an
investigation into whether this is feasible.

This results in about 11% speed-up of self-play.