Randomise the opening during self-play games
Implement a temperature parameter that is applied to the eight first
moves during self-play. The purpose of this parameter is to encourage
exploration and ensure diversity in the self-play games without straying
too far from what the tree search suggest.

The temperature is applied during move selection, after the tree search
has finished, by raising each visit count to the inverse of the
temperature. This means that when a low temperature is given the
selection acts greedily, while a high temperature results in a uniform
random selection between the moves.

This temperature is also applied to the GTP client, but with a much
smaller value to avoid it picking some of the exploratory moves
that we allow during self-play. This does result in a slight drop in
strength when playing against previous version, using the same weights,
as we sometimes play sub-optimal moves.