Re-use search tree between successive moves
During self-play and in the GTP client, re-use the sub-tree of the
previous search tree that corresponds to the selected and opponents
move. This avoids the search tree having to re-discover information that
is learned higher up in the search tree.

This provides a significant boost to performance, going from 35% to 50%
win rate against commit `cafeb18e`.