Add a transposition table to the monte carlo search tree
Store the result of each neural network computation so that those
results can be re-used if the same board position occurs again. This is
useful because, for example, the following moves results in the same
board position and computing the neural network twice for it is
wasteful:

* `d4 q16 d16`
* `d16 q16 d4`

This is extra true because the transposition table is re-used between
different self-play instances, where especially the first few moves of
each game is mostly the same. Some simple trials shows about a 5%
performance improvements for five self-play games.