Fix the transposition table to take super-ko into account
Include the path used to get to the board position into the
transposition table. This negates some of the advantages in using
t-table but prevents the engine from playing invalid moves, besides when
a ko is active the order of the moves matters **a lot**.

One will still benefit massively from using the t-table using repeated
self-play, and it still provides about a 3% speed-up compared to without
a t-table.