Allow the user to control the number of samples to extract per file
Add the environment variable `NUM_SAMPLES` that can be used the number
of samples the data extractor will pluck from each game. This value can
be given as either an absolute number or a percentage.

This patch is mostly in preparation for the reinforcement learning as we
will need to pluck more than the one sample per game we used during
supervised learning.