Allow for >100% number of samples in the data extractor
Add the option to specify that more than 100% of all board positions
should be extracted using the `NUM_SAMPLES` environment variables.

The way this works is that one copy of each board position is created
for every valid symmetry, this augmented the data set is then sampled
from. This data set can be up to 800% larger (depending on how many
board position symmetries are rejected because they are the same as
the identity) than the original.