Yield inside of the spinlock
Yield to the OS scheduler inside of the spinlock implementation to make
it nicer to CPU on shared system.

This gives about a 5% speed-up, presumably because we run 64 threads by
default, which is more than then number of cores on the benchmarking
machine (and most machines in general). More importantly it is 35% more
efficient in terms CPU usage.