Re-factor how feature symmetries are computed
Previously we always computed the features for the identity symmetry and
then transforming the computed features into whatever symmetry we
desired. This is a bit wasteful since we always know what the symmetry
to compute is before the features are computed anyway, so just write the
features in the correct symmetry immediately to avoid having to
re-shuffle all of the memory.

Also do some minor re-factoring of symmetries:

* Use unsafe `get_unchecked` in the `apply` method to get better
  performance
* Rename `count_zero` to `count_zeros` (and some cosmetics)
* Store symmetry lookup table as `u16` instead of `usize` to reduce
  memory usage.