Re-write some hot code as AVX2
Change the UCT calculations, and the argmax implementation in the
`probe` code path into assembly avx2 to gain an approximate 10x speed-up
for those specific methods. This gives an 1.5% speed-up during self-play
but more importantly reduces the cpu time by 11%.

The RAVE implementation of `Value` is still not converted to avx2 since
it is not in active use, but there is nothing hard that would stop us
from re-writing it.