Add tool for scoring undecided games
Add the script `tools/sgf2score.py` that reads SGF games from standard
input and if there is no winner, it uses GNU Go to determine who won and
write that as the result to standard output.

This is useful to avoid having to implement a full chinese scoring
algorithm into _Dream Go_. Especially since most of the self-play games
ends with a Resign anyway (based on the last self-play session of 1,800
games only 36 games ended without a resignation).