Fix bug in the GTP interface when the opponent turn was skipped
When no action was given for the opponent, and instead two
consecutive calls to `genmove` were made the GTP client failed to
fast-forward the search tree. Causing the wrong color search tree
to be used, this failed an assertion further in the code.

This patch always fast-forward the search tree when generating new
moves if the color does not match.