Add benchmark for the PUCT and RAVE value functions
Fix some weird numeric instability where for some reason the following
two were not equal:

1. `a * (1 / b)`
2. `a / b`

Based on Intel documentation the reciprocal should have a maximum error
of 1e-12, which should be small enough to not be a problem. But it
causes errors in the range of 1e-1.

The benchmarks also show that the AVX2 implementation are between eight
to 12 times better.