Add support for Tromp-Taylor scoring
Implement scoring in the `go` module and implement the `final_score`
command in the GTP interface.

This is useful because the planned policy-only self-play games will not
know when to stop, and we need some way to determine the score outside
of resignation (which in the main termination criteria in mcts self-play
games).