Add a specialized function for self-play using only the policy
Add the `--policy-play` command-line option that plays self-play games
using only the prior value in the root node and performs **no search**.
This procedure is parallelized over multiple games instead of within the
search to get a better GPU utilization.

Simple benchmarks suggests this is about 3x faster than just running
with `NUM_ITER=1` for a `BATCH_SIZE=16`. This is lower than the expected
5x performance improvement expected from the batch size benchmarks.