Make the GTP interface write a summary of the search tree to stderr
Once a move has been generated from the GTP interface, write a summary
of the search tree to standard error for easy debugging and as a "nice
interface" for human to understand what the engine looked at in general.

The exact formatting of the summary is heavily influences by Leela.