Add the value property to each node during self-play
During self-play also add the value property `V` to each node that
contains how good the situation is (according to the monte carglo search
tree) for black.

Also forbid resigning in 5% of all self-play games so that we can get a
dataset of complete games to determine the value to resign at. An
additional advantage of this is that the engine will see end game and
pass moves instead of just resignations all the time.