Change the parallel self-play model
Instead of playing only one self-play game at a time, but running many
parallel probes into the search tree. During self-play play multiple
games in addition to parallel probes.

This is better because when the search tree is empty, or during end-game
when there are not that many paths to probe the current parallel model
does not fully saturate our batch size. This new model will however
always achieve full saturation because we:

* Play `batch_size` number of games in parallel, and there is always at
  least one _valid_ probe per game.
* Perform multiple neural network evaluations in parallel, resulting in
  a lower average response time.
* Less contention between threads probing into the same search tree, so
  the CPU time spent spinning on the locks is smaller.

Benchmarks suggests an 18% improvement when playing five self-play
games.