Normalize the value function Q(s,a) to the range [0,1)
The value function Q(s,a) in both PUCT and RAVE represents the
probability of the current player winning the game for the given state
and action.

Previously we used the value from the _value head_ from the neural
network directly as the value function. But this is wrong since the
value head is in the range [-1,+1] instead of the desired [0,1) by the
value function. This patch fixes this by normalizing the _value head_ to
the correct range.

This yield a non-trivial strength improvement, with the fixed version of
the program winning `18 - 2` over the previous version after the
MCTS constants were re-tuned.