Remove batch normalization operators during inference
Since during inference the batch normalization operator just acts as a
fixed linear transformation we can merge it into the surrounding
convolutional layers instead of having it be its own separate layer.

This gives a speed-up of about 1.23x, and should also give improved
precision since this merge is done pre-storing as `f16` which causes
most of the loss in precision.