Add support for data extraction using the ExIt scheme
When the data extractor encounters a SGF with only the correct move
given for any one position (as it would be in a normal SGF file), it
would emit a _partial_ policy with just that move. This patch adds the
command line switch `--exit` that makes the data extractor perform a
full search for every partial policy it encounters, and then emits the
result of the search.

This scheme is what is known as an Expert Iteration (ExIt), and is the
scheme suggested by "Thinking Fast and Slow with Deep Learning and
Tree Search". This is approximately 250 times faster than the AlphaZero
approach to self-play but will generate lower quality data.