Fix passing moves in the monte carlo search tree
Make the monte carlo tree search aware of passing moves, and score them
appropriately according to the Tromp-Taylor rules. This makes the engine
aware of the exact score at the end of the game, and by-pass the neural
network evaluation (this might be a problem due to the supervised games
we bootstrap with not playing with Tromp-Taylor rules, but we will see).

This also fixes a serious bug where the prior value for the passing move
suggested by the neural network was thrown away, making the engine never
pass.