Re-design the neural network package
In order to support DP4A we needed to re-design the neural network
package to support storing scaling factors for each tensor. This
was not straight forward using the old design, so this is an almost full
rewrite of the neural network code to be more modular.

This also includes some other augmentations, such as dynamic algorithm
selection, tensor core support for cuBLAS and a more flexible weights
format.

There is no performance improvement in this patch outside of DP4A which
still needs tweaks to get the correct scale at everything.