Introduce `Service` which allows us to clean up some background threads
In order to clean-up the background threads, and fix some
synchronization bugs, introduce the `Service` model that allows one to
build a service that process and respond to requests using background
threads in a standalized way.

This patch also includes some changes to how the configuration is
handled by removing the `Param` trait from the `mcts` package and moving
it into `Config` as `lazy_static` variables instead.