Fix race condition in `WorkspaceGuard`
Fix a race condition that could cause a single `Workspace` to be owned
by two `WorkspaceGuard`. This is because we were using an `Rc` to wrap
the `Workspace`, which would then be dropped when the `WorkspaceGuard`
is dropped (but the ownership is handed over in the `drop` function,
which is called before the struct is dropped).