Re-write the services to ensure the requests are processed in-order
In order to ensure that the request that has `has_more` set to false is
executed *last* re-write the services to keep the services state in a
mutex by itself, and tie the `has_more` value to that mutex.

This might be slightly slower in theory since it involes more locking,
but should be trivial in practice.