Add docker images for running the database and GPU worker
Some initial images for running the database and GPU workers necessary
to run _Dream Go_ in a distributed fashion.

The database in particular is so far pretty trivial and we may want to
invest in something better than SQLite3 if the load gets higher or we
require more reliable storage.