Reduce the number of residual blocks from 19 to 9
In order to make self-play feasible reduce the number of residual blocks
from 19 to 9, this provides roughly a 2x speed-up with minimal (?)
strength decrease. Some other minor tweaks has been done to the network
and search in this patch:

- Add support for FPU (disabled)
- Add support for virtual loss count larger than 1.
- Add ladder features to the neural network
- Add _vertex is only reachable from one color_ features to the neural
  network.
- Enable biases for the convolutional layers in the residual blocks.

When all of these changes are combined the 9x128 neural network seems to
be roughly equal to the 19x128 neural network after training on strong
expert games.

The playoff below is performed using only the policy network of either
program as the search in the two programs works somewhat differently now
and it is hard to get fair comparison:

---

playoff: gomill

dg-v050p v dg-minip (100/100 games)
board size: 19   komi: 7.5
           wins              black         white       avg cpu
dg-v050p     51 51.00%       20 40.00%     31 62.00%      1.88
dg-minip     49 49.00%       19 38.00%     30 60.00%      1.52
                             39 39.00%     61 61.00%