Re-factor the `go` package
The `go` package was turning into a bit of a mess because of the
speculative playout that are necessary to read ladders. To fix this I
broke the `Board` struct into two separature structs `BoardFast` which
contains only a minimal board implementation and `Board` which contains
all of the additional meta-information.

This split allows for speculative playout with less of a mess. Also
introduce some macros to help with very common operations.

There are no performance of strength implications of this change.