Use a more compact bitset representation for training examples
Instead of the previous, bloated, representation for examples where each
element was a half precision number, we now use a bitset representation
where each element in the features is represented by a single bit, the
same applies to the value and the policy is quantized to one byte per
element.

This yield a reduction of about 13.5x in example size, which is
significant since we were running into space issues when training on
large datasets.