Apply the MCTS temperature during policy play too
A lot of policy play games were decided based on which player were
unlucky a picked a _noise move_. This is not good, as we would prefer
the neural network picking moves that are somewhat strongly recommended
by the network.

This patch applies the `config::TEMPERATURE` constant that is used
during MCTS to discourage bad moves too the policy play too. This seems
to have a large effect as the _noise moves_ are largely gone while
preserving the random nature of the policy play games.