Score with a rule based playout instead of flood-filling
To better be able to determine the life-and-death of large dragons
change the scoring algorithm to use a playout based on a simple rule
based policy that works in two phases:

1. Attempt to secure all current groups.
2. Attempt to capture opponent groups.

Once these two converge we score using the previous flood-filling
algorithm. This works pretty well but it has issues with:

- Seki
- Nakade eye shapes (both living and killing)

Over all it brings down the misclassification from 2.7% to 1.9% over
1,000 samples, and more importantly the misclassifications are less
systematic (with the exception of seki), so the network should learn to
ignore the outliers rather then learn the incorrect scoring algorithm we
used before.