Re-score games before feeding them to the training process
Do not rely on our internal scoring algorithm, instead use GNU Go to
score any (finished) games before feeding them to training process.

We do not score resigned games since GNU Go is only about 5 kyu in
strength, which is not enough to accurately score games that we are not
finished. But it's seki and end-game detection is pretty good.