Parallel scoring of games
Score multiple games in parallel instead of sequentially. This is useful
because GNU Go scoring of games ends up being the bottleneck of the
initial stage, and since GNU Go only use a single thread we may as well
run multiple instances at once.

This has the side-effect of shuffling the games, since the threading
model is non-deterministic.