Fix compilation on nightly rust
There seems to be some bug with the latest nightly bug where it does not
recognize temporary backward labels in inline assembly code. It
interprete `jnz 1b` as jump to byte value instead of jump to the
backward label `1`.

This patch replaces all backward lables with explicit labels and forbid
inlining (to avoid duplicate labels). This is hopefully a temporary
workaround.