Increase the temperature during policy play
Change how the temperature works during policy play, before we used
constant temperature during the entire run. But this is bad for several
reasons:

- Playing away during the middle of an important fight will scew the
  result a lot, resulting in a worse value head.
- We do not get a very high amount of noise during the opening causing
  the network to get stuck playing 4-4 and 4-3 points.

To avoid the pitfalls of the first point we previously used a fairly low
temperature, but this meant we got a lot of very similar games. Now we
instead start with a _high_ temperature and then decrease it over the
course of the game. This should avoid both of the problems mentioned
above.