Add a parallel dilation to each residual layer
Replace the first convolution of each residual layer with two
convolutions with different dilations. So each residual block now
looks like:

```
    x
  / | \
  | A  B
  | | /
  | C
  \ |
    y
```

Where `A` is a normal 3x3 convolution, and `B` is a 3x3 convolution with
a dilation of 2. This should help the neural network see wider, and has
been observed by lightvector [1] that this help with for example large
dead dragons. `C` is a normal 3x3 convolution.

[1] https://github.com/lightvector/GoNN#dilated-convolutions-mar-2018