Implement `kgs-genmove_cleanup`
Add a version of `genmove` that does not allow for passing, except when
the board is fully scoreable according to the TT-rules.

This command is used on the KGS server when the two parties cannot
decide on a winner, and they need to play the rest of the game out.