Implement the `final_status_list` and re-vamp the scoring heuristic
Instead of using a rule-based heuristic for scoring the game, which did
not work **too** well (only about 95% of all games were correctly
scored), use the neural network to perform the rollout. This is slower
but a lot more precise, and as an added benefit it should improve as the
neural network learns more.

Also implement the `final_status_list` GTP command that returns the
vertices on the board that are of status:

- alive
- dead
- seki (not supported)