Optimize CUDA memory copying for `HtoD` and `DtoH`
Since each call to `cudaMemcpyAsync` has a fairly large constant factor
associated with it, reduce the number of calls to this function by
making sure the features are already continuous in host memory.

This involves a bit of juggling of our typeless arrays so that they can
be extended and split off in a performant manner.

The main impact of this patch is that large batch sizes (256) can be up
to 5% faster, while more modest batch sizes (16) are 2.5% faster. This
also shifts the GPU profile a bit as especially the policy head was
suffering badly from this overhead since each output was so small. The
percentage of time spent in each part of the neural network is now
roughly:

Input: 2.3055144%
Upsample: 1.9010382%
Residual Blocks: 94.10813%
Policy Head: 0.3100984%
Value Head: 0.323581%

This does not sum to 100% so I must have lost a percentage somewhere,
but you get the general idea of where to put the optimization effort.

Further improvements might be to pre-copy the features to the GPU, which
could get ride of almost the entire _Input_ block.