Re-write the neural network to use quantized weights
Perform a complete re-write of the cuDNN neural network code in order to
support inference using INT8. This is significant because:

- It gives about a 3.5x performance boost for all batch sizes.
- It limits support to NVIDIA Pascal (and later) GPU's
- It limits us to certain network architectures, most notable the cuDNN
  kernels for INT8 only supports dilation = 1 (which prevents us from
  implementing certain enchancements suggested by @lightvector)

There are also a bunch of minor / trivial changes that were made during
the re-write that gives minor performance benefits, such as:

- Multi GPU support
- Improved performance of CUDA events (by not tracking timing
  information)
- Use of pinned memory for copying of network weights to devices

The tensorflow training code was also re-factored to make use of
fp16*, this allows us to double the batch size with no observed
performance loss.

* Note that older GPUs will use pseudo-half, which has the performance
  of fp32 but the accuracy of fp16. Newer GPUs (Volta+) will use Tensor
  Cores, which provides a significant performance boost.