Extend the lifetime of all slots to the full duration of an inference
There was a race condition in the neural network code, where the
following two lifetime requirements were not in-sync:

- The lifetime of a `Slot`.
- The _lifetime_ of an async CUDA call.

By extending the lifetime of all `Slots` to the entire duration of an
inference pass we can avoid this issue. Which prevents the same memory
location from being used at the same time by different threads.