Use a proper adversarial network for color balance
Change the adversarial network used for color balancing to use a more
standard training mechanic, where the discriminator gets 3x the training
of the main network. This is to ensure it is sufficiently strong to
detect any bias.

This unfortunately has significant effect on the training performance,
dropping from `0.75` steps per second to `0.32`.