Always use the strongest network when generating self-play games
Change the self-play procedure to always use the strongest network when
generating games. This has a few implications on the rest of the chain
too, since we need to ensure we get timely evaluations of each network,
and that we do not traing while pending network evaluation.

This should help stabilize the self-play.