Add support for komi using the color planes
Add support for setting komi to values other than `7.5`. This is
accomplished by using the colour planes as suggested in #25.

There are some additional changes included in this changelist because
setting the colour planes to non-binary values causes our feature
representation, which is based on bit sets, to break-down. Hence I also
re-wrote the feature to TFRecord's.

\#25 https://github.com/Chicoryn/dream-go/issues/25#issuecomment-381417656