Fix a rare crash in the symmetry elimination code
Fix a rare crash during symmetry elimination that could happen when a
board is symmetric in multiple ways, the symmetric points forms a chain,
and the chain is not traversed in the correct order. For example
consider the following symmetric points:

```
A <-> B
B <-> C
```

If we following the chains, in-order, the we will transfer the value of
`A` to `B`, then the value of `B` to `C`. However if we try it in the
opposite order, by first transferring `B` to `C`, and then try
transferring the value of `A` to `B`. This will fail since we have
already invalidated `B`.

We solve this by keeping track of where each value has moved, and for
each value we want to move follow the non-identity edges.