Remove the --extract option
Remove the --extract option as a valid command line option in favour of
proper support for adding policy, and value, vectors to both self-play
and policy play games.

Getting ride of .tfrecord format from the binary allows us to drop a lot
of complexity, and makes it easier to visualize the training (since we
now store the policy together with the board state).