Reduce the size of `BoardFast` from 1090 bytes to 736 bytes
In order to improve performance of the SGF parser, which puts most of
its effort into copying the board states around, reduce the size of the
internal board representation from 1090 bytes to 736 bytes by packing its
fields into a bitstruct.

This results in a speed-up between 2.5% to 18.7% during SGF parsing.