Bias towards uncertain moves when deciding which move to search with `--ex-it`
When deciding on what move to search for `--ex-it` do so based on the
skewness of the policy for each move. Weighting them such that moves
with a high skewness are less likely to be picked, because a search on
those moves should not have time to explore anything anyway.