Add support for gradient checkpoints, and other improvements
Add support for gradient checkpoints to the training script. This allows
us to use **a lot** less memory during training for a given batch
size `n`, hence we can increase the batch size by a lot. The new default
batch size is not 2048.

Also make the following changes to the training procedure:

- Add gradient boosting, based on how unique a board position is. So
  duplicate board positions, mostly from the beginning of the game, will
  now have a smaller effect on the gradient.
- Normalize convolution filters per output channel instead of globally.
- Add an orthogonal loss to the TensorBoard metrics to track.