Use the NHWC data layout during training
With the advent of tensor cores it seems like it is more efficient to
perform convolutions in NHWC than NCHW. So change the data layout we use
in the training code.

There are still some casts from NHWC to NCHW in the internal Tensorflow
code, but this seems to yield about a 7% performance improvement. It
also unifies the data layout used during training and inference, which
should result in less future work.