Do not allow the engine to play _worthless_ moves even with `kgs-genmove_cleanup`
Change the clean up mode so that if the engine passes, and the game is
still not scorable according to the Tromp-Taylor moves, then it force
the engine to play in one of the areas that does not count as territory
for either player.

This fixes the problem where the engine, when using clean up, would
waste a lot of moves filling in extra eyes and generally playing in an
unenjoyable fashion.

Remaining problems:
- How to handle seki