Do not allow the ponder service to exceed the specified number of rollouts
If the user has given a rollout limit with `--num-rollout`, do not allow
that number to be exceeded while pondering. However if no rollout limit
is given, allow pondering with a minimum tree size of 500,000.

The purpose of this change is to allow more parallel processing during
tournaments with given rollout limits. Since previously the only way to
keep it fair was to turn off pondering, but that resulted in poor GPU
utilization sometimes when just one engine could not fill up the GPU.