Try to preserve the ponder state across `clear_board` and `komi`
Only throw away the ponder state when the board state actually
changes, instead of at every `clear_board` command. This should improve
the engines strength in the opening for CGOS-style tournaments where
the engine is left _spinning_ for a long time before the game begins.

These changes should still conform to the GTP specification, but could
be considered a bit dirty since the engine can effectively use more than
the total game time. However, it is in the spirit of pondering.