Add support for deterministic training
Add a `--deterministic` flag to the training script, which set the
seed of all random operations to a fixed value. This is very useful
when verifying models, since it ensures we check against the same data
every time.

Also add a `--mask` argument, which takes a `;` delimited list. It will
multiple the features by the given mask before training / evaluating
on it.